Move MessageReader loop from MessageDispatcher to ProtocolEndpoint
This change moves the loop that reads messages from the MessageReader
from the MessageDispatcher to ProtocolEndpoint.  This allows the
MessageDispatcher to be used separately from a specific connection so
that message handlers can be registered and shared between connections.
